ec: add new script for emacsclient
To remind myself later:

I got rid of my old script because all it did was handle the logic for
finding the server file socket (now handled by $EMACS_SERVER_FILE) and
for tramp integration (now handled by $EMACSCLIENT_TRAMP).

Which is great! In theory, everything should now use 'emacsclient' and
things should automatically work.

There are a few edge-cases though:

*) Hammerspoon. Apparently, due to bugs in its lua, we need to spawn
   from a script to not have Hammerspoon crash. Ugh.

But there are a few advantages to the script:

*) we can now pipe to emacs which is handy

*) open a file at a certain line (future)