hammerspoon: fix stale emacs server file
If Hammerspoon happened to start before emacs did, then it would get the
wrong value for the location of this file and be forever wrong (until
hammerspoon reloaded).

It's fast enough to read the location of this file every time we shell
out to emacsclient so we just do that instead.