doom: fix logical error in ansible spawning a process
Previously, we added an empty string for a missing argument. Either that
never worked or `make-process' changed or ansible changed to read that
empty string argument as a file name, resulting in an error that ""
could not be found.

Most likely, this never worked. Either way, this fixes the logic to add
`nil' instead of an empty string and then later filter out all nulls.