doom: remove pyvenv logic around switching workspaces
This causes a massive slowdown in emacs startup time due to force
loading lots of packages and then checking the disk for python
environments.

Luckily, we don't really use this feature anymore so just go back to
manually activating environments when needed.

Also, for added history, this feature was added before we used lsp which
has its own workflow now.